requirements.txt
build
flake8
numpy
oresat-olaf>=2.2.0
